+++ src/app/beer_garden/router.py
@@ -290,14 +290,16 @@ def update_api_heartbeat(operation: Operation):
         operation.source_garden_name is not None
         and operation.source_garden_name != config.get("garden.name")
         and operation.source_api is not None
+        and operation.operation_type == "PUBLISH_EVENT"
+        and operation.model.name == Events.GARDEN_SYNC.name
     ):
         source_garden = getattr(gardens, operation.source_garden_name, None)
-
-        beer_garden.garden.check_garden_receiving_heartbeat(
-            operation.source_api,
-            garden_name=operation.source_garden_name,
-            garden=source_garden,
-        )
+        if operation.model.payload.name == operation.source_garden_name:
+            beer_garden.garden.check_garden_receiving_heartbeat(
+                operation.source_api,
+                garden_name=operation.source_garden_name,
+                garden=source_garden,
+            )
 
 
 def invalid_source_check(operation: Operation):
@@ -697,6 +699,7 @@ def _pre_forward(operation: Operation) -> Operation:
         operation.model.has_parent = False
         operation.model.source_garden = None
         operation.model.target_garden = None
+        operation.model.metadata = {}
 
         # Pull out and store the wait event, if it exists
         wait_event = operation.kwargs.pop("wait_event", None)